Add sound test buttons for Pomodoro settings
