+++ src/locales/de/translation.json
@@ -92,6 +92,7 @@
     "customThemeName": "Name des Themes",
     "addTheme": "Theme hinzufügen",
     "addSound": "Sound hinzufügen",
+    "testSound": "Sound testen",
     "customSoundUrl": "Sound-URL",
     "yourSounds": "Eigene Sounds",
     "yourThemes": "Eigene Themes",
