+++ src/locales/en/translation.json
@@ -92,6 +92,7 @@
     "customThemeName": "Theme name",
     "addTheme": "Add Theme",
     "addSound": "Add Sound",
+    "testSound": "Test Sound",
     "customSoundUrl": "Sound URL",
     "yourSounds": "Your Sounds",
     "yourThemes": "Your Themes",
